Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main(){
     int b;
     int c;
     scanf("%d %d %d",&a ,&b ,&c);
-    int avg = (a+b+c)/3;
-    print("Average: %d",avg);
+    float avg = (a+b+c)/3;
+    print("Average: %.2f",avg);
     return 0;
 }
